[ABStock.Agents] implement MarketMakerAgent

+++ src/ABStock.Agents/Models/AgentState.cs
@@ -7,7 +7,7 @@ public class AgentState
     public string AgentName { get; set; } = string.Empty;
     public AgentType AgentType { get; set; }
     public decimal Cash { get; set; }
-    public int Position { get; set; }
+    public decimal Position { get; set; }
     
     public decimal GetPortfolioValue(decimal currentPrice) => Cash + Position * currentPrice;
 }
